rename counselor template
This is to better differentiate between the two templates.
